+++ components/src/index.js
@@ -9,6 +9,7 @@ export { Button, PrimaryButton, DangerButton, QuietButton, IconicButton } from "
 export { Link } from "./Link";
 export { InlineValidation, HeaderValidation } from "./Validation";
 
+export { Alert } from "./Alert";
 export { Branding, BrandingText } from "./Branding";
 export { DropdownMenu, DropdownLink, DropdownButton, DropdownItem } from "./DropdownMenu";
 export { HelpText, RequirementText, FieldLabel, MaybeFieldLabel } from "./FieldLabel";